Added Team Logos to non-logged in pages
- Added team logos to non-logged in pages
  -> Stats, Sched, Results, etc
- Added basic default image
- Changed game page to use new default image
- Added Bilingual about for Teams & Venues
- Fixed some c:out issues
- Added league name to team page
- Changed weather around to store modified data
  -> Wind & Gusts stored in km/h
  -> Pressure now a double and correct in db
  -> Removed calc from index page
- Changed wind to not show any decimal points
- Updated db to reflect changes
  -> Added stored procedure

+++ LastEver/src/beans/WeatherBean.java
@@ -15,7 +15,7 @@ public class WeatherBean {
 	private String weatherIcon;
 	private int weatherCode;
 	private String weatherDescription;
-	private int weatherPressure;
+	private Double weatherPressure;
 	private int weatherHumidity;
 	private Double weatherWind;
 	private Double weatherGust;
@@ -51,7 +51,7 @@ public class WeatherBean {
 		return this.weatherDescription;
 	}
 	
-	public int getWeatherPressure() {
+	public Double getWeatherPressure() {
 		return this.weatherPressure;
 	}
 
@@ -101,7 +101,7 @@ public class WeatherBean {
 		this.weatherDescription = this.weatherDescription.replaceAll(" ", "_");
 	}
 	
-	public void setWeatherPressure(int wp) {
+	public void setWeatherPressure(Double wp) {
 		this.weatherPressure = wp;
 	}
